move all relevant files into package folder
