--- .eslintignore
@@ -1,2 +1,0 @@
-dist/
-app/choreo-modeler/node_modules
