+++ app/choreo-modeler/.gitignore
@@ -1,0 +1,5 @@
+.DS_Store
+node_modules
+dist
+*.log
+.idea/
