+++ app/choreo-modeler/.travis.yml
@@ -9,4 +9,3 @@ cache:
 script:
   - npm run lint
   - npm run test
-  - npm run bundle
