+++ app/choreo-modeler/package.json
@@ -1,21 +1,24 @@
 {
-  "name": "choreo-js",
-  "version": "1.0.0",
-  "description": "A choreography diagram rendering toolkit and web modeler.",
+  "name": "chor-js",
+  "version": "0.0.1",
+  "description": "A BPMN 2.0 choreography diagram rendering toolkit and web modeler.",
   "keywords": [
     "choreography model",
-    "distributed ledger technology",
+    "choreography",
     "modeler",
     "browser",
     "choreo-js",
-    "toolkit"
+    "toolkit",
+    "bpmn"
   ],
   "main": "index.js",
   "directories": {
     "lib": "lib"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "all": "run-s lint test",
+    "lint": "eslint .",
+    "test": "karma start"
   },
   "author": {
     "name": "Jan Ladleif",
@@ -24,12 +27,35 @@
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/jan-ladleif/choreography-modeler"
+    "url": "https://github.com/jan-ladleif/chor-js"
+  },
+  "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babelify": "^8.0.0",
+    "browserify": "^16.1.0",
+    "chai": "^4.1.2",
+    "eslint": "^4.18.1",
+    "eslint-plugin-bpmn-io": "^0.4.1",
+    "karma": "^2.0.0",
+    "karma-browserify": "^5.2.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-firefox-launcher": "^1.1.0",
+    "karma-ie-launcher": "^1.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-safari-launcher": "^1.0.0",
+    "karma-spec-reporter": "0.0.26",
+    "mocha": "^5.0.1",
+    "mocha-test-container-support": "^0.2.0",
+    "stringify": "^3.1.0",
+    "puppeteer": "~1.10.0"
   },
   "dependencies": {
     "bpmn-js": "^3.0.2",
     "diagram-js": "^3.0.1",
     "min-dash": "^3.2.0",
-    "tiny-svg": "^2.2.0"
+    "tiny-svg": "^2.2.0",
+    "inherits": "^2.0.1"
   }
 }
