+++ app/choreo-modeler/test/resources/subChoreographies.png
