+++ app/choreo-modeler/test/resources/tasksWithLoopType.bpmn
