+++ app/choreo-modeler/test/resources/tasksWithLoopType.png
