+++ app/choreo-modeler/test/resources/tasksWithMultiplicities.png
